Added error checking

+++ fileFunctions.py
@@ -1,11 +1,9 @@
 # FILE:     fileFunctions.py
 # AUTHOR:   David LeBlanc 143807L@acadiau.ca
 # DATE:     2019/11/17
-# VERSION:  1.0
+# VERSION:  1.1
 # PURPOSE:  This program defines functions for use in managing the files needed
 # for the Chadatonic.
-
-# Test
 
 import sys
 import tkinter as tk
@@ -32,7 +30,19 @@ def create_directories():
 
 
 # Imports a selected sound into the sounds folder so it is easy to find
+# Returns -1 on error, 0 on success
 def import_sound(soundPath):
     ext = str(os.path.splitext(soundPath)[1])
     path = str(Path(soundPath).stem)
+
+    # Return error if unusable format
+    if ext != ".wav":
+        return -1
+
+    # Return error if directories do not exist
+    if not os.path.exists("sounds"):
+        return -1
+
     copyfile(soundPath, "sounds/" + path + ext)
+
+    return 0
